Added part parent type selection/saving
also rolled back some of snave's changes (sorry snave) because i'm a JS noob and got confused
