+++ README.md
@@ -1,7 +1,3 @@
 # Figua Blockbench Plugin
 uwu
 
-## How to setup development environment
-- `yarn` - Build the project
-- `yarn build:scripts` - initial setup command for custom build scripts.
-- `yarn watch` - Starts watching the src for changes, and builds the project when changes are found
