+++ package.json
@@ -1,38 +1,18 @@
 {
-  "name": "java-animator",
+  "name": "blockbench-plugins",
   "version": "1.0.0",
-  "description": "",
+  "description": "Blockbench plugins repository.",
   "main": "index.js",
-  "scripts": {
-    "setup": "npm install && npm run build:scripts",
-    "go": "npm install && npm run build:scripts && echo for future runs use npm run watch && npm run watch",
-    "build:scripts": "./node_modules/.bin/rollup --config ./config/tools.rollup.config.js",
-    "watch": "node dist/tools/scripts/watch.js",
-    "build:app": "node dist/tools/scripts/build.js",
-    "rollup-debug": "./node_modules/.bin/rollup --config ./config/dev.animated-java.rollup.config.js",
-    "build": "npm run build:scripts && npm run build:app",
-    "dev": "npm run watch"
+  "types": "types/index.d.ts",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/JannisX11/blockbench-plugins.git"
   },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@babel/core": "^7.12.3",
-    "@babel/plugin-external-helpers": "^7.12.1",
-    "@babel/plugin-proposal-class-properties": "^7.12.1",
-    "@babel/preset-env": "^7.12.1",
-    "@babel/preset-react": "^7.12.1",
-    "@rollup/plugin-node-resolve": "^9.0.0",
-    "@rollup/plugin-url": "^5.0.1",
-    "@rollup/plugin-yaml": "^3.0.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-latest": "^6.24.1",
-    "chalk": "^4.1.0",
-    "exists-command": "^0.1.2",
-    "rollup": "^2.32.1",
-    "rollup-plugin-babel": "^4.4.0",
-    "rollup-plugin-commonjs": "^10.1.0",
-    "rollup-plugin-replace": "^2.2.0",
-    "rollup-plugin-terser": "^7.0.2"
+  "bugs": {
+    "url": "https://github.com/JannisX11/blockbench-plugins/issues"
+  },
+  "homepage": "https://github.com/JannisX11/blockbench-plugins#readme",
+  "devDependencies": {
+    "blockbench-types": "^3.8.0"
   }
 }
