+++ schema.json
@@ -1,0 +1,38 @@
+{
+	"$id": "https://blockbench.net/plugins.schema.json",
+	"$schema": "http://json-schema.org/draft-07/schema#",
+	"type": "object",
+	"propertyNames": {
+		"pattern": "^[a-z0-9._-]+$"
+	},
+	"additionalProperties": {
+		"type": "object",
+		"required": ["title", "author", "icon", "description"],
+		"properties": {
+			"title": {
+				"type": "string"
+			},
+			"icon": {
+				"type": "string"
+			},
+			"author": {
+				"type": "string"
+			},
+			"description": {
+				"type": "string"
+			},
+			"about": {
+				"type": "string"
+			},
+			"version": {
+				"type": "string"
+			},
+			"variant": {
+				"type": "string"
+			},
+			"min_version": {
+				"type": "string"
+			}
+		}
+	}
+}
