--- src/figura-plugin.js
@@ -1,3 +1,0 @@
-// / <reference path="./types/index.d.ts" />
-
-import "./plugin-definition"
